[evol] add a not-so-clean way to retrieve information on a phone (vendor, model and MAC address) from its IP address by querying the provisioning server.
Note that if you want to be able to get these informations from a phone that has not yet been provisioned, you need to install the 'sofia-sip-bin' package. This is because if the server can not find the phone information from its DB, it tries sending a SIP OPTIONS to the phone, and it does this by calling an external command ('sip-options') instead of manipulating its own SIP stack. This was simpler.


git-svn-id: https://rcs.lan.proformatique.com/svn/xivo/trunk@9338 dc85baef-7018-0410-bfbd-8a639da648bb