[evol] SIP NOTIFY/SCCP reset synchronize commands can now be proxied through one or more asterisk via AMI.
Before, for some phones supporting synchronize via SIP NOTIFY, it was sent
directly by the provisioning server. This was causing some problems:
- if the phone was using a TCP or TLS transport, the phone would not listen
  for the notify
- also, there was an issue of knowing which user the phone was currently
  registered as. This is because we don't keep the last synchronized config
  in memory. We don't do this because of the overhead of knowing when the
  device is synchronized or not (but we might do it someday).
- the usual firewall configuration issue

Also, it was not possible to synchronize SCCP devices this way.

Patches for asterisk were needed because of the way sip peers are registered
and to add more options for autocreated peers...


git-svn-id: https://rcs.lan.proformatique.com/svn/xivo/trunk@11005 dc85baef-7018-0410-bfbd-8a639da648bb