[evol] new xivo-aastra-3.2.2.56 plugin.
Remove old xivo-aastra-3.2.1.1019 plugin.

Also some refactoring. Now takes into account the major version number
in automatic plugin association, so that a device with a major version
that is not the same as the major version of the firmware will get a
slitghly lower score.

Remove some not so useful comments.