add a global NAT option
When the option is set, it's possible to have more than 1 device with
the same IP address.

More precisely, when a HTTP/TFTP request is processed by the provisining
server, the server won't clear all the 'ip' key of all the other devices
with the same IP address.

This also means the xivo-provd process is more efficient when the NAT
option is activated because it skips the "find and update all the
devices with that IP address" step, which is something that happens often
when lots of device are behind a NAT, especially if they are rebooted
all at the same time.