autocreate: modify the factory to reuse the same username
with pjsip we can use the anonymous endpoint and aor to allow unauthenticated
devices to be registered.

this change will allow the autocreation process of new peers to inherit the
username of the base config without adding a suffix to its username.

making the username of all devices in autoprov "anonymous"

I also change the id to be a uuid instead of being generated from the current